split up common
