+++ src/common/constants.ts
@@ -1,0 +1,5 @@
+export const API_ENDPOINT = 'https://runfission.com'
+
+export default {
+  API_ENDPOINT
+}
